ft_memmove.c and norminette

+++ ft_strlen.c
@@ -6,16 +6,17 @@
 /*   By: rthome-d <rthome-d@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/11 21:21:40 by rthome-d          #+#    #+#             */
-/*   Updated: 2022/05/12 17:50:34 by rthome-d         ###   ########.fr       */
+/*   Updated: 2022/05/26 15:52:42 by rthome-d         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
-	while(s[i])
+	while (s[i])
 		i++;
 	return (i);
 }
